apollo-link-batch-http: disallow GET with batches
There is no spec for batching over GET. I don't believe apollo-server supports
it given that there's no spec. We can always come up with a spec and implement
it later.
