+++ packages/apollo-link-http/src/__tests__/sharedHttpTests.ts
@@ -2,7 +2,6 @@ import { Observable, ApolloLink, execute } from 'apollo-link';
 import { print } from 'graphql';
 import gql from 'graphql-tag';
 import * as fetchMock from 'fetch-mock';
-import objectToQuery from 'object-to-querystring';
 
 const sampleQuery = gql`
   query SampleQuery {
@@ -497,53 +496,6 @@ export const sharedHttpTest = (
           expect(signal).toBe('foo');
           expect(mode).toBe('no-cors');
           expect(headers['content-type']).toBe('application/json');
-        }),
-      );
-    });
-
-    it('supports using a GET request', done => {
-      const variables = { params: 'stub' };
-
-      let requestedString;
-      const customFetch = (uri, options) => {
-        const { body, ...newOptions } = options;
-        const queryString = objectToQuery(convertBatchedBody(body));
-        requestedString = uri + queryString;
-        return fetch(requestedString, newOptions);
-      };
-      const link = createLink({
-        uri: 'data',
-        fetchOptions: { method: 'GET' },
-        fetch: customFetch,
-      });
-
-      execute(link, { query: sampleQuery, variables }).subscribe(
-        makeCallback(done, result => {
-          const [uri, options] = fetchMock.lastCall();
-          const { method, body, ...rest } = options;
-          expect(body).toBeUndefined();
-
-          expect(method).toBe('GET');
-        }),
-      );
-    });
-
-    it('supports using a GET request on the context', done => {
-      const variables = { params: 'stub' };
-      const link = createLink({
-        uri: 'data',
-      });
-
-      execute(link, {
-        query: sampleQuery,
-        variables,
-        context: {
-          fetchOptions: { method: 'GET' },
-        },
-      }).subscribe(
-        makeCallback(done, result => {
-          const method = fetchMock.lastCall()[1].method;
-          expect(method).toBe('GET');
         }),
       );
     });
